backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
polars>=1.0.0
openpyxl>=3.1.0
pdfplumber>=0.11.0
python-docx>=1.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
redis>=5.0.0
celery[redis]>=5.4.0
python-multipart>=0.0.9
pydantic-settings>=2.3.0
aiofiles>=24.1.0
